[level 0] Title: 간단한 식 계산하기, Time: 0.79 ms, Memory: 47.4 MB -BaekjoonHub

+++ C#/프로그래머스/0/181865. 간단한 식 계산하기/간단한 식 계산하기.cs
@@ -4,14 +4,14 @@ public class Solution
 {
     public int solution(string binomial) 
     {
-        int answer = 0;
+        long answer = 0;
         string[] op = binomial.Split(' ');
-        int a = int.Parse(op[0]);
-        int b = int.Parse(op[2]);
+        long a = int.Parse(op[0]);
+        long b = int.Parse(op[2]);
         if(op[1] == "+") answer = a + b;
         else if(op[1] == "-") answer = a - b;
         else if(op[1] == "*") answer = a * b;
         else answer = a / b;
-        return answer;
+        return (int)answer;
     }
 }
